feat(docs): Deprecate to `dorssel/usbipd-win`
Remove docs based on `cezuni/usbip-win` and replace with a recommendation for and links
to [dorssel/usbipd-win](https://github.com/dorssel/usbipd-win).

The `dorssel/usbipd-win` package is more complete and polished, coming very close to "it
just works":

- Installable OOTB with a package manager: `> winget install -y -e usbipd-win`
- Installs a proper Windows service for `> usbipd`
- WSL2 convenience commands and docs for end-to-end set up
- Docs are quite complete and clear

Getting up and running took me a fraction of the time of my attempts to get it all
working with `cezuni/usbip-win`.  I only discovered `dorssel/usbipd-win` when I started
trying to figure out how to get `cezuni/usbip-win`'s `> usbipd` working as a service.
It would have saved me a bunch of time if I'd discovered `dorssel/usbipd-win` first so
this change is made in hopes of saving others time as well.

Apologies if this ruffled any feathers.  Feel free to reject and close this PR, it will
at least be google'able for others who will be better served by the approach
`dorssel/usbipd-win`.